refactor: move helpers into separate file + deprecate inject mode
