+++ lib/runtime/component-normalizer.js
@@ -5,7 +5,7 @@
 // be included in the final webpack user bundle.
 
 export default function normalizeComponent (
-  rawScriptExports,
+  scriptExports,
   compiledTemplate,
   functionalTemplate,
   injectStyles,
@@ -13,14 +13,12 @@ export default function normalizeComponent (
   moduleIdentifier, /* server only */
   shadowMode /* vue-cli only */
 ) {
-  var esModule
-  var scriptExports = rawScriptExports = rawScriptExports || {}
+  scriptExports = scriptExports || {}
 
   // ES6 modules interop
-  var type = typeof rawScriptExports.default
+  var type = typeof scriptExports.default
   if (type === 'object' || type === 'function') {
-    esModule = rawScriptExports
-    scriptExports = rawScriptExports.default
+    scriptExports = scriptExports.default
   }
 
   // Vue.extend constructor export interop
@@ -103,7 +101,6 @@ export default function normalizeComponent (
   }
 
   return {
-    esModule: esModule,
     exports: scriptExports,
     options: options
   }
