+++ test.html
@@ -1,0 +1,1 @@
+<script src="./output.js"></script>
