--- test/fixtures/inject.js
@@ -1,1 +1,0 @@
-window.injector = require('!!../../lib/loader?inject!./inject.vue')
