--- test/fixtures/inject.vue
@@ -1,15 +1,0 @@
-<template>
-  <div class="msg">{{ msg }}</div>
-</template>
-
-<script>
-const SomeService = require('./service')
-
-module.exports = {
-  data () {
-    return {
-      msg: SomeService.msg
-    }
-  }
-}
-</script>
